Release 23

+++ certoscale/src/main/java/com/certoclav/certoscale/util/FTPManager.java
@@ -4,7 +4,11 @@ package com.certoclav.certoscale.util;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
+import android.util.Log;
+import android.widget.Toast;
 
+import com.certoclav.certoscale.database.Protocol;
+import com.certoclav.certoscale.supervisor.ApplicationManager;
 import com.certoclav.library.application.ApplicationController;
 
 import org.apache.commons.net.ftp.FTPClient;
@@ -15,9 +19,17 @@ import org.apache.commons.net.ftp.FTPReply;
 import java.io.File;
 import java.io.FileFilter;
 import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
+import es.dmoral.toasty.Toasty;
 import needle.Needle;
 
 //import com.adeel.library.easyFTP;
@@ -93,7 +105,7 @@ public class FTPManager {
         try {
             if (folderILIMS.charAt(0) != '/')
                 folderILIMS = "/" + folderILIMS;
-        }catch (Exception e){
+        } catch (Exception e) {
             folderILIMS = "/IFP_ILIMS";
             e.printStackTrace();
         }
@@ -101,7 +113,7 @@ public class FTPManager {
         try {
             if (folder.charAt(0) != '/')
                 folder = "/" + folder;
-        }catch (Exception e){
+        } catch (Exception e) {
             folder = "/IFP";
             e.printStackTrace();
         }
@@ -123,18 +135,16 @@ public class FTPManager {
         editor.putString(FOLDER_ILIMS, folderILIMS);
         editor.putBoolean(FTP_STATUS, isEnabled);
         editor.commit();
-
-        updateAll(null);
     }
-/*
-*  Uploading to other folder read upload from the folder if the uploading is success remove from the folder.
-*   There are two folder
-*   IFP_UPLOADING are used for raw data
-*   IFP_ILIMS_UPLOADING are used for ash data for iLims
- */
+    /*
+     *  Uploading to other folder read upload from the folder if the uploading is success remove from the folder.
+     *   There are two folder
+     *   IFP_UPLOADING are used for raw data
+     *   IFP_ILIMS_UPLOADING are used for ash data for iLims
+     */
 
-    public void updateAll(final FTPListener listener) {
-        if(!isEnabled)return;
+    public void uploadProtocols(final FTPListener listener, final List<Protocol> protocols, final boolean isForce) {
+        if (!isEnabled) return;
         Needle.onBackgroundThread().execute(new Runnable() {
                                                 @Override
                                                 public void run() {
@@ -152,46 +162,64 @@ public class FTPManager {
                                                                 ftp.setControlEncoding("UTF-8");
                                                                 ftp.setAutodetectUTF8(true);
                                                                 ftp.makeDirectory(folder);
-                                                                ftp.changeWorkingDirectory(folder);
-                                                                FTPFile[] files = ftp.listFiles(folder, filter);
-                                                                File dir = new File(android.os.Environment.getExternalStorageDirectory(), "IFP_ILIMS/RAW_DATA");
-                                                                dir.mkdirs();
+                                                                if (!ftp.changeWorkingDirectory(folder)) {
+                                                                    if (listener != null)
+                                                                        listener.onUploading(false, "Es gibt ein Problem mit " + folder + " Ordner!");
+                                                                    return;
+                                                                }
                                                                 ftp.setFileType(org.apache.commons.net.ftp.FTP.BINARY_FILE_TYPE);
 
-                                                                for (File file : dir.listFiles(csvFilter)) {
+                                                                boolean isUploaded;
+                                                                for (Protocol protocol : protocols) {
+                                                                    protocol.generateCSV();
+                                                                    isUploaded = protocol.isUploaded();
+                                                                    if ((!isForce && isUploaded) || protocol.getIsPending())
+                                                                        continue;
+                                                                    isUploaded = false;
+                                                                    //Uploading not uploaded protocols
+                                                                    File file = createTempProtocolFile(protocol);
+                                                                    if (file == null)
+                                                                        continue;
                                                                     FileInputStream srcFileStream = new FileInputStream(file);
-                                                                    if(file.getName().contains("-deletemenow")){
-                                                                        ftp.deleteFile(file.getName().replace("-deletemenow",""));
+                                                                    ftp.deleteFile(file.getName());
+                                                                    if (ftp.storeFile(file.getName(), srcFileStream)) {
                                                                         file.delete();
-                                                                        srcFileStream.close();
+                                                                        isUploaded = true;
+                                                                    } else {
+                                                                        if (listener != null)
+                                                                            listener.onUploading(false, null);
+                                                                    }
+                                                                    srcFileStream.close();
+
+
+                                                                    //Uploading iLIMS protocols
+                                                                    ftp.makeDirectory(folderILIMS);
+                                                                    ftp.changeWorkingDirectory(folderILIMS);
+                                                                    file = createTempProtocolFileForIlims(protocol);
+                                                                    if (file == null) {
+                                                                        Log.e("nuuuuuuuuuul", "nuuuuuuuuuuuuul");
                                                                         continue;
                                                                     }
+                                                                    srcFileStream = new FileInputStream(file);
                                                                     ftp.deleteFile(file.getName());
-                                                                    ftp.storeFile(file.getName(), srcFileStream);
-                                                                    srcFileStream.close();
-                                                                }
-                                                                dir = new File(dir, "/IFP_UPLOADING");
-                                                                for (File file : dir.listFiles(csvFilter)) {
-                                                                    FileInputStream srcFileStream = new FileInputStream(file);
-                                                                    ftp.deleteFile(file.getName());
-                                                                    if (ftp.storeFile(file.getName(), srcFileStream))
+                                                                    if (ftp.storeFile(file.getName(), srcFileStream)) {
                                                                         file.delete();
-                                                                    srcFileStream.close();
-                                                                }
-
-                                                                dir = new File(new File(android.os.Environment.getExternalStorageDirectory(), "IFP_ILIMS/RAW_DATA"),
-                                                                        "/IFP_ILIMS_UPLOADING");
-                                                                ftp.makeDirectory(folderILIMS);
-                                                                ftp.changeWorkingDirectory(folderILIMS);
-                                                                for (File file : dir.listFiles(csvFilter)) {
-                                                                    FileInputStream srcFileStream = new FileInputStream(file);
-                                                                    if(ftp.storeFile(file.getName(), srcFileStream)){
-                                                                        file.delete();
+                                                                    } else {
+                                                                        isUploaded = false;
+                                                                        if (listener != null)
+                                                                            listener.onUploading(false, null);
                                                                     }
                                                                     srcFileStream.close();
+
+                                                                    if (isUploaded) {
+                                                                        if (listener != null)
+                                                                            listener.onUploaded(protocol);
+                                                                    } else {
+                                                                        if (listener != null)
+                                                                            listener.onUploading(false, "Das Hochladen ist fehlgeschlagen!");
+                                                                    }
                                                                 }
-                                                                if (listener != null)
-                                                                    listener.onUpdated();
+
                                                             } catch (Exception e) {
                                                                 if (listener != null)
                                                                     listener.onUploading(false, e.getMessage());
@@ -210,6 +238,127 @@ public class FTPManager {
                                             }
         );
     }
+
+    private File createTempProtocolFile(Protocol protocol) {
+
+        String data = protocol.generateCSV();
+        File file = null;
+
+        String[] contents = protocol.getContent().split("\n");
+        String date = null;
+        if (contents!=null && contents.length > 2) {
+            Log.d("content", contents[1]);
+            date = contents[1];
+            date = date.replaceAll("-", "").replaceAll(":", "").trim().replaceAll(" ", "_");
+        }
+        if(date==null){
+            DateFormat df = new SimpleDateFormat("yyyyMMdd_HHmmss");
+            date = df.format(new Date(System.currentTimeMillis()));
+        }
+//28 May 2019 08:56:50 GMT
+//        SimpleDateFormat dateFormat = new SimpleDateFormat("dd MMM yyyy HH:mm:ss", Locale.ENGLISH);
+//        dateFormat.setTimeZone(TimeZone.getTimeZone("GMT"));
+//        Date protocolDate = null;
+//        try {
+//            protocolDate = dateFormat.parse(protocol.getDate());
+//        } catch (ParseException e) {
+//            e.printStackTrace();
+//        }
+//        Log.d("date", protocol.getDate() + " " + df.format(protocolDate));
+
+
+        String protocolName = protocol.getAshSampleName()
+                + "_" + date;
+
+        file = new File(ApplicationController.getContext().getCacheDir(), protocolName + ".csv");
+
+
+        if (file == null)
+            return null;
+        try {
+            FileOutputStream f = new FileOutputStream(file);
+            PrintWriter pw = new PrintWriter(f);
+            pw.print(data);
+            pw.flush();
+            pw.close();
+            f.close();
+            return file;
+
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+            android.util.Log.e("ExportUtils", "******* File not found. Did you" +
+                    " add a WRITE_EXTERNAL_STORAGE permission to the   manifest?");
+            Toasty.error(ApplicationController.getContext(), "Can't write to storage", Toast.LENGTH_SHORT, true).show();
+            return null;
+        } catch (IOException e) {
+            e.printStackTrace();
+            Log.e("Export Utils", "io exception: " + e.toString());
+            return null;
+        }
+    }
+
+    private File createTempProtocolFileForIlims(Protocol protocol) {
+
+        protocol.generateCSV();
+        File file = null;
+
+        String[] contents = protocol.getContent().split("\n");
+        String date = null;
+        if (contents!=null && contents.length > 2) {
+            Log.d("content", contents[1]);
+            date = contents[1];
+            date = date.replaceAll("-", "").replaceAll(":", "").trim().replaceAll(" ", "_");
+        }
+        if(date==null){
+            DateFormat df = new SimpleDateFormat("yyyyMMdd_HHmmss");
+            date = df.format(new Date(System.currentTimeMillis()));
+        }
+
+
+        String protocolName = protocol.getAshSampleName()
+                + "_" + date;
+        file = new File(ApplicationController.getContext().getCacheDir(), protocolName + ".csv");
+
+
+        if (file == null) {
+            Log.e("3", "333333333333");
+            return null;
+        }
+        try {
+            FileOutputStream f = new FileOutputStream(file);
+            PrintWriter pw = new PrintWriter(f);
+
+            StringBuilder sb = new StringBuilder();
+            sb.append("samplenumber" + "," + protocol.getAshSampleName() + "\r\n");
+            sb.append("sampleweight" + "," + ApplicationManager.getInstance().getTransformedWeightAsString(
+                    protocol.getSampleWeight())
+                    .replaceAll(",", ".") + "\r\n");
+            sb.append("ashweight" + "," + ApplicationManager.getInstance().getTransformedWeightAsString(
+                    protocol.getLastAshWeight()).replaceAll(",", ".") + "\r\n");
+            sb.append("ashpercent" + "," + protocol
+                    .getAshResultPercentageAsString().replaceAll(",", ".") + "\r\n");
+
+            String data = sb.toString();
+            pw.print(data);
+            pw.flush();
+            pw.close();
+            f.close();
+            return file;
+
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+            android.util.Log.e("ExportUtils", "******* File not found. Did you" +
+                    " add a WRITE_EXTERNAL_STORAGE permission to the   manifest?");
+            Toasty.error(ApplicationController.getContext(), "Can't write to storage",
+                    Toast.LENGTH_SHORT, true).show();
+            return null;
+        } catch (IOException e) {
+            e.printStackTrace();
+            Log.e("Export Utils", "io exception: " + e.toString());
+            return null;
+        }
+    }
+
 
     private List<CustomFile> findNotUploadedFiles(FTPFile[] inFtp, File[] inLocal) {
 
@@ -269,12 +418,12 @@ public class FTPManager {
         password = preferences.getString(PASSWORD, "admin");
         folder = preferences.getString(FOLDER, "/IFP");
         folderILIMS = preferences.getString(FOLDER_ILIMS, "/IFP_ILIMS");
-        isEnabled = preferences.getBoolean(FTP_STATUS,true);
+        isEnabled = preferences.getBoolean(FTP_STATUS, true);
 
         try {
             if (folderILIMS.charAt(0) != '/')
                 folderILIMS = "/" + folderILIMS;
-        }catch (Exception e){
+        } catch (Exception e) {
             folderILIMS = "/IFP_ILIMS";
             e.printStackTrace();
         }
@@ -282,7 +431,7 @@ public class FTPManager {
         try {
             if (folder.charAt(0) != '/')
                 folder = "/" + folder;
-        }catch (Exception e){
+        } catch (Exception e) {
             folder = "/IFP";
             e.printStackTrace();
         }
@@ -291,7 +440,7 @@ public class FTPManager {
     public static interface FTPListener {
         void onConnection(boolean isConnected, String message);
 
-        void onUpdated();
+        void onUploaded(Protocol protocol);
 
         void onUploading(boolean isUploaded, String message);
     }
